Added CreateObject(regionhandle, userID, itemID) to post objects that are to be fetched from the user's inventory server and rezzed in the region. Added all code necessary to fetch the item and the asset, and rez it inworld. The access to the item is uncap-ed and unverified -- I may place it later either under a cap or with auth verification. But in this model regions don't have the user's inventory, so they would have to guess the item IDs. Added safemode config to Standalone Hypergrid, similar effect to AllowRegionAccessToInventory in Inventory Server. Everyone should have these vars set to their default values except me!

+++ OpenSim/Region/CoreModules/Hypergrid/HGStandaloneInventoryModule.cs
@@ -54,6 +54,7 @@ namespace OpenSim.Region.CoreModules.Hypergrid
         private static readonly ILog m_log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
         private static bool initialized = false;
         private static bool enabled = false;
+        private static bool safemode = false;
 
         private bool m_doLookup = false;
         Scene m_scene;
@@ -77,6 +78,8 @@ namespace OpenSim.Region.CoreModules.Hypergrid
 
                 // This module is only on for standalones
                 enabled = !config.Configs["Startup"].GetBoolean("gridmode", true) && config.Configs["Startup"].GetBoolean("hypergrid", false);
+                if (config.Configs["Hypergrid"] != null)
+                    safemode = config.Configs["Hypergrid"].GetBoolean("safemode", false);
             }
         }
 
@@ -116,34 +119,35 @@ namespace OpenSim.Region.CoreModules.Hypergrid
 
         public virtual void AddHttpHandlers(IHttpServer httpServer)
         {
+            if (!safemode)
+            {
+                httpServer.AddStreamHandler(
+                    new RestDeserialiseSecureHandler<Guid, InventoryCollection>(
+                        "POST", "/GetInventory/", m_inventoryService.GetUserInventory, CheckAuthSession));
+                httpServer.AddStreamHandler(
+                    new RestDeserialiseSecureHandler<InventoryItemBase, bool>(
+                        "POST", "/DeleteItem/", m_inventoryBase.DeleteItem, CheckAuthSession));
+                httpServer.AddStreamHandler(
+                    new RestDeserialiseSecureHandler<InventoryFolderBase, bool>(
+                        "POST", "/UpdateFolder/", m_inventoryBase.UpdateFolder, CheckAuthSession));
 
-            httpServer.AddStreamHandler(
-                new RestDeserialiseSecureHandler<Guid, InventoryCollection>(
-                    "POST", "/GetInventory/", m_inventoryService.GetUserInventory, CheckAuthSession));
+                httpServer.AddStreamHandler(
+                    new RestDeserialiseSecureHandler<InventoryFolderBase, bool>(
+                        "POST", "/MoveFolder/", m_inventoryBase.MoveFolder, CheckAuthSession));
+
+                httpServer.AddStreamHandler(
+                    new RestDeserialiseSecureHandler<InventoryFolderBase, bool>(
+                        "POST", "/PurgeFolder/", m_inventoryBase.PurgeFolder, CheckAuthSession));
+            }
 
             httpServer.AddStreamHandler(
                 new RestDeserialiseSecureHandler<InventoryFolderBase, bool>(
                     "POST", "/NewFolder/", m_inventoryBase.AddFolder, CheckAuthSession));
 
             httpServer.AddStreamHandler(
-                new RestDeserialiseSecureHandler<InventoryFolderBase, bool>(
-                    "POST", "/UpdateFolder/", m_inventoryBase.UpdateFolder, CheckAuthSession));
-
-            httpServer.AddStreamHandler(
-                new RestDeserialiseSecureHandler<InventoryFolderBase, bool>(
-                    "POST", "/MoveFolder/", m_inventoryBase.MoveFolder, CheckAuthSession));
-
-            httpServer.AddStreamHandler(
-                new RestDeserialiseSecureHandler<InventoryFolderBase, bool>(
-                    "POST", "/PurgeFolder/", m_inventoryBase.PurgeFolder, CheckAuthSession));
-
-            httpServer.AddStreamHandler(
                 new RestDeserialiseSecureHandler<InventoryItemBase, bool>(
                     "POST", "/NewItem/", m_inventoryBase.AddItem, CheckAuthSession));
 
-            httpServer.AddStreamHandler(
-                new RestDeserialiseSecureHandler<InventoryItemBase, bool>(
-                    "POST", "/DeleteItem/", m_inventoryBase.DeleteItem, CheckAuthSession));
 
         }
